+++ src/resources/projects/index.ts
@@ -13,6 +13,7 @@ export {
   type ProjectRetrieveResponse,
   type ProjectUpdateResponse,
   type ProjectListResponse,
+  type ProjectRetrieveParams,
   type ProjectUpdateParams,
   type ProjectListParams,
 } from './projects';