+++ tests/index.test.ts
@@ -24,6 +24,7 @@ describe('instantiate client', () => {
       baseURL: 'http://localhost:5000/',
       defaultHeaders: { 'X-My-Default-Header': '2' },
       apiKey: 'My API Key',
+      project: 'example-project',
     });
 
     test('they are used in the request', () => {
@@ -87,14 +88,19 @@ describe('instantiate client', () => {
         error: jest.fn(),
       };
 
-      const client = new StainlessV0({ logger: logger, logLevel: 'debug', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        logger: logger,
+        logLevel: 'debug',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
 
       await forceAPIResponseForClient(client);
       expect(debugMock).toHaveBeenCalled();
     });
 
     test('default logLevel is warn', async () => {
-      const client = new StainlessV0({ apiKey: 'My API Key' });
+      const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
       expect(client.logLevel).toBe('warn');
     });
 
@@ -107,7 +113,12 @@ describe('instantiate client', () => {
         error: jest.fn(),
       };
 
-      const client = new StainlessV0({ logger: logger, logLevel: 'info', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        logger: logger,
+        logLevel: 'info',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
 
       await forceAPIResponseForClient(client);
       expect(debugMock).not.toHaveBeenCalled();
@@ -123,7 +134,7 @@ describe('instantiate client', () => {
       };
 
       process.env['STAINLESS_V0_LOG'] = 'debug';
-      const client = new StainlessV0({ logger: logger, apiKey: 'My API Key' });
+      const client = new StainlessV0({ logger: logger, apiKey: 'My API Key', project: 'example-project' });
       expect(client.logLevel).toBe('debug');
 
       await forceAPIResponseForClient(client);
@@ -140,7 +151,7 @@ describe('instantiate client', () => {
       };
 
       process.env['STAINLESS_V0_LOG'] = 'not a log level';
-      const client = new StainlessV0({ logger: logger, apiKey: 'My API Key' });
+      const client = new StainlessV0({ logger: logger, apiKey: 'My API Key', project: 'example-project' });
       expect(client.logLevel).toBe('warn');
       expect(warnMock).toHaveBeenCalledWith(
         'process.env[\'STAINLESS_V0_LOG\'] was set to "not a log level", expected one of ["off","error","warn","info","debug"]',
@@ -157,7 +168,12 @@ describe('instantiate client', () => {
       };
 
       process.env['STAINLESS_V0_LOG'] = 'debug';
-      const client = new StainlessV0({ logger: logger, logLevel: 'off', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        logger: logger,
+        logLevel: 'off',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
 
       await forceAPIResponseForClient(client);
       expect(debugMock).not.toHaveBeenCalled();
@@ -173,7 +189,12 @@ describe('instantiate client', () => {
       };
 
       process.env['STAINLESS_V0_LOG'] = 'not a log level';
-      const client = new StainlessV0({ logger: logger, logLevel: 'debug', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        logger: logger,
+        logLevel: 'debug',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
       expect(client.logLevel).toBe('debug');
       expect(warnMock).not.toHaveBeenCalled();
     });
@@ -185,6 +206,7 @@ describe('instantiate client', () => {
         baseURL: 'http://localhost:5000/',
         defaultQuery: { apiVersion: 'foo' },
         apiKey: 'My API Key',
+        project: 'example-project',
       });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/foo?apiVersion=foo');
     });
@@ -194,6 +216,7 @@ describe('instantiate client', () => {
         baseURL: 'http://localhost:5000/',
         defaultQuery: { apiVersion: 'foo', hello: 'world' },
         apiKey: 'My API Key',
+        project: 'example-project',
       });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/foo?apiVersion=foo&hello=world');
     });
@@ -203,6 +226,7 @@ describe('instantiate client', () => {
         baseURL: 'http://localhost:5000/',
         defaultQuery: { hello: 'world' },
         apiKey: 'My API Key',
+        project: 'example-project',
       });
       expect(client.buildURL('/foo', { hello: undefined })).toEqual('http://localhost:5000/foo');
     });
@@ -212,6 +236,7 @@ describe('instantiate client', () => {
     const client = new StainlessV0({
       baseURL: 'http://localhost:5000/',
       apiKey: 'My API Key',
+      project: 'example-project',
       fetch: (url) => {
         return Promise.resolve(
           new Response(JSON.stringify({ url, custom: true }), {
@@ -230,6 +255,7 @@ describe('instantiate client', () => {
     const client = new StainlessV0({
       baseURL: 'http://localhost:5000/',
       apiKey: 'My API Key',
+      project: 'example-project',
       fetch: defaultFetch,
     });
   });
@@ -238,6 +264,7 @@ describe('instantiate client', () => {
     const client = new StainlessV0({
       baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
       apiKey: 'My API Key',
+      project: 'example-project',
       fetch: (...args) => {
         return new Promise((resolve, reject) =>
           setTimeout(
@@ -270,6 +297,7 @@ describe('instantiate client', () => {
     const client = new StainlessV0({
       baseURL: 'http://localhost:5000/',
       apiKey: 'My API Key',
+      project: 'example-project',
       fetch: testFetch,
     });
 
@@ -279,12 +307,20 @@ describe('instantiate client', () => {
 
   describe('baseUrl', () => {
     test('trailing slash', () => {
-      const client = new StainlessV0({ baseURL: 'http://localhost:5000/custom/path/', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        baseURL: 'http://localhost:5000/custom/path/',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/custom/path/foo');
     });
 
     test('no trailing slash', () => {
-      const client = new StainlessV0({ baseURL: 'http://localhost:5000/custom/path', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        baseURL: 'http://localhost:5000/custom/path',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/custom/path/foo');
     });
 
@@ -293,35 +329,39 @@ describe('instantiate client', () => {
     });
 
     test('explicit option', () => {
-      const client = new StainlessV0({ baseURL: 'https://example.com', apiKey: 'My API Key' });
+      const client = new StainlessV0({
+        baseURL: 'https://example.com',
+        apiKey: 'My API Key',
+        project: 'example-project',
+      });
       expect(client.baseURL).toEqual('https://example.com');
     });
 
     test('env variable', () => {
       process.env['STAINLESS_V0_BASE_URL'] = 'https://example.com/from_env';
-      const client = new StainlessV0({ apiKey: 'My API Key' });
+      const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
       expect(client.baseURL).toEqual('https://example.com/from_env');
     });
 
     test('empty env variable', () => {
       process.env['STAINLESS_V0_BASE_URL'] = ''; // empty
-      const client = new StainlessV0({ apiKey: 'My API Key' });
+      const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
       expect(client.baseURL).toEqual('https://api.stainless.com');
     });
 
     test('blank env variable', () => {
       process.env['STAINLESS_V0_BASE_URL'] = '  '; // blank
-      const client = new StainlessV0({ apiKey: 'My API Key' });
+      const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
       expect(client.baseURL).toEqual('https://api.stainless.com');
     });
   });
 
   test('maxRetries option is correctly set', () => {
-    const client = new StainlessV0({ maxRetries: 4, apiKey: 'My API Key' });
+    const client = new StainlessV0({ maxRetries: 4, apiKey: 'My API Key', project: 'example-project' });
     expect(client.maxRetries).toEqual(4);
 
     // default
-    const client2 = new StainlessV0({ apiKey: 'My API Key' });
+    const client2 = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
     expect(client2.maxRetries).toEqual(2);
   });
 
@@ -331,6 +371,7 @@ describe('instantiate client', () => {
         baseURL: 'http://localhost:5000/',
         maxRetries: 3,
         apiKey: 'My API Key',
+        project: 'example-project',
       });
 
       const newClient = client.withOptions({
@@ -357,6 +398,7 @@ describe('instantiate client', () => {
         defaultHeaders: { 'X-Test-Header': 'test-value' },
         defaultQuery: { 'test-param': 'test-value' },
         apiKey: 'My API Key',
+        project: 'example-project',
       });
 
       const newClient = client.withOptions({
@@ -375,6 +417,7 @@ describe('instantiate client', () => {
         baseURL: 'http://localhost:5000/',
         timeout: 1000,
         apiKey: 'My API Key',
+        project: 'example-project',
       });
 
       // Modify the client properties directly after creation
@@ -404,20 +447,22 @@ describe('instantiate client', () => {
   test('with environment variable arguments', () => {
     // set options via env var
     process.env['STAINLESS_V0_API_KEY'] = 'My API Key';
-    const client = new StainlessV0();
+    const client = new StainlessV0({ project: 'example-project' });
     expect(client.apiKey).toBe('My API Key');
+    expect(client.project).toBe('example-project');
   });
 
   test('with overridden environment variable arguments', () => {
     // set options via env var
     process.env['STAINLESS_V0_API_KEY'] = 'another My API Key';
-    const client = new StainlessV0({ apiKey: 'My API Key' });
+    const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
     expect(client.apiKey).toBe('My API Key');
+    expect(client.project).toBe('example-project');
   });
 });
 
 describe('request building', () => {
-  const client = new StainlessV0({ apiKey: 'My API Key' });
+  const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
 
   describe('custom headers', () => {
     test('handles undefined', () => {
@@ -436,7 +481,7 @@ describe('request building', () => {
 });
 
 describe('default encoder', () => {
-  const client = new StainlessV0({ apiKey: 'My API Key' });
+  const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project' });
 
   class Serializable {
     toJSON() {
@@ -521,7 +566,12 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new StainlessV0({ apiKey: 'My API Key', timeout: 10, fetch: testFetch });
+    const client = new StainlessV0({
+      apiKey: 'My API Key',
+      project: 'example-project',
+      timeout: 10,
+      fetch: testFetch,
+    });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
     expect(count).toEqual(2);
@@ -551,7 +601,12 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new StainlessV0({ apiKey: 'My API Key', fetch: testFetch, maxRetries: 4 });
+    const client = new StainlessV0({
+      apiKey: 'My API Key',
+      project: 'example-project',
+      fetch: testFetch,
+      maxRetries: 4,
+    });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
 
@@ -575,7 +630,12 @@ describe('retries', () => {
       capturedRequest = init;
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
-    const client = new StainlessV0({ apiKey: 'My API Key', fetch: testFetch, maxRetries: 4 });
+    const client = new StainlessV0({
+      apiKey: 'My API Key',
+      project: 'example-project',
+      fetch: testFetch,
+      maxRetries: 4,
+    });
 
     expect(
       await client.request({
@@ -606,6 +666,7 @@ describe('retries', () => {
     };
     const client = new StainlessV0({
       apiKey: 'My API Key',
+      project: 'example-project',
       fetch: testFetch,
       maxRetries: 4,
       defaultHeaders: { 'X-Stainless-Retry-Count': null },
@@ -637,7 +698,12 @@ describe('retries', () => {
       capturedRequest = init;
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
-    const client = new StainlessV0({ apiKey: 'My API Key', fetch: testFetch, maxRetries: 4 });
+    const client = new StainlessV0({
+      apiKey: 'My API Key',
+      project: 'example-project',
+      fetch: testFetch,
+      maxRetries: 4,
+    });
 
     expect(
       await client.request({
@@ -667,7 +733,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new StainlessV0({ apiKey: 'My API Key', fetch: testFetch });
+    const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project', fetch: testFetch });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
     expect(count).toEqual(2);
@@ -697,7 +763,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new StainlessV0({ apiKey: 'My API Key', fetch: testFetch });
+    const client = new StainlessV0({ apiKey: 'My API Key', project: 'example-project', fetch: testFetch });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
     expect(count).toEqual(2);